requirements.txt
pytest
python-pptx